Lomonosov tablebases
7-man tablebases were calculated in 2012 on Computer Science department of Moscow State University. Calculations were done on supercomputer named Lomonosov. So they were named Lomonosov tablebases. Generation of tablebases is an important step as in chess theory and practice so in improving of retrograde analysis algorithm.

The tablebases contain the exact evaluations (draw or moves to mate) for all positions with no more than 7 pieces on the board. The total number of positions in the tables – more than 500 trillions (500 000 billions). The real number of similar positions is several times more, but part of them can be obtained from others by mirroring and rotating of the board, so there is no need to keep them on disks.